googleアカウントでのログイン機能実装

+++ db/migrate/20241123130354_remove_name_from_trees.rb
@@ -1,5 +1,7 @@
 class RemoveNameFromTrees < ActiveRecord::Migration[7.1]
   def change
-    remove_column :trees, :name, :string
+    if column_exists?(:trees, :name)
+      remove_column :trees, :name, :string
+    end
   end
 end
